elster/requirements.txt
arxiv==1.4.7
feedparser==6.0.10
jellyfish==0.11.2
numpy==1.24.3
psycopg2-binary==2.9.7
pypdf==3.9.0
PyPika==0.48.9
sgmllib3k==1.0.0
tqdm==4.65.0
